requirements.txt
Flask==2.1.2
selenium==4.2.0
line-bot-sdk==2.2.1
webdriver-manager==3.7.0
gunicorn==20.1.0
pycryptodome==3.14.1